refactor(core): move the hook types and knowlHome to core, where both sides can reach them
Two declarations were filed with one of their consumers rather than beneath
all of them, and each produced a dependency cycle.

NormalizedHostHook / HookHost sat in cli/agents/host-hook.ts beside the
function that produces one. The session layer consumes them, so four store
modules imported upward into cli/. host-hook.ts re-exports them, so every
existing import site is unchanged.

knowlHome() sat in workspace/paths.ts, but it is machine-local state and has
nothing to do with multi-repo workspaces -- the model cache, the repo
registry and the resume store all resolve against it. core/startup-trace.ts
importing it out of workspace/ was the tell: core is supposed to be the
bottom.

No behaviour change. 1967 tests pass, typecheck clean.
