+++ src/cli/agents/host-hook.ts
@@ -1,62 +1,24 @@
 import path from 'node:path';
 import { validateKnowledgeWrite } from '../../core/knowledge-validation.js';
 import { SessionEventType } from '../../core/types.js';
+import type {
+  HookHost, NormalizedHookEventName, NormalizedHostHook,
+} from '../../core/host-hook-types.js';
 import { isSessionEventType } from './lifecycle.js';
 import { hostProfile, isHookHost } from './hosts/index.js';
 import { AgentName } from './types.js';
 
-export type HookHost = 'codex' | 'claude' | 'cursor' | 'claude-desktop' | 'generic';
-export type NormalizedHookEventName =
-  | 'session-start'
-  | 'turn-start'
-  | 'session-event'
-  | 'checkpoint'
-  | 'turn-stop'
-  | 'session-stop'
-  | 'agent-start'
-  | 'agent-stop';
-
-export interface NormalizedHostHook {
-  host: HookHost;
-  event: NormalizedHookEventName;
-  externalSessionId: string;
-  externalTurnId?: string;
-  projectRoot: string;
-  title?: string;
-  status?: 'finished' | 'failed';
-  type?: SessionEventType;
-  payload: Record<string, unknown>;
-  /** True when this tool event is a Knowl MCP/CLI call — used to reset the drift reminder. */
-  knowlTool?: boolean;
-  /**
-   * The Knowl tool this event fired for, normalized to its bare name.
-   *
-   * Hosts prefix MCP tools (`mcp__knowl__knowl_store`), so the raw name cannot be matched
-   * against what the MCP server recorded under its own tool name.
-   */
-  knowlToolName?: string;
-  /**
-   * What makes this event distinct from its neighbours, for debounce purposes only.
-   *
-   * Never displayed or stored. The capture fingerprint is otherwise built from the
-   * payload, and a non-shell tool event carries only `summary: "<Tool> completed"` --
-   * so two different Grep calls looked identical and the second inside the debounce
-   * window was dropped silently, taking its change card with it. The debounce still
-   * needs to collapse the same event delivered twice, so this keys on the tool's
-   * arguments rather than on a timestamp or a counter.
-   */
-  captureKey?: string;
-  /** Claude subagent id. Present on every subagent event, absent on main-thread events. */
-  agentId?: string;
-  /** Claude subagent type, e.g. "Explore". Used only to title the binding. */
-  agentType?: string;
-  /**
-   * Titles and ids the caller supplied in its own tool_input, used to recognise
-   * this agent's own writes in new commits. Held in memory for comparison only and
-   * never persisted, so no attribution column is needed.
-   */
-  knowlChangeKeys?: { ids: string[]; titles: string[] };
-}
+/**
+ * The hook types are declared in `core/host-hook-types.ts` and re-exported here.
+ *
+ * They moved because the session layer consumes them and would otherwise have to import
+ * upward into `cli/`, making the hook path a dependency cycle. The re-export keeps every
+ * existing `from '.../host-hook.js'` import working — normalizing a hook and describing
+ * the result are still one idea from a caller's point of view.
+ */
+export type {
+  HookHost, NormalizedHookEventName, NormalizedHostHook,
+} from '../../core/host-hook-types.js';
 
 const MAX_STRING = 2_000;
 const MAX_RETAINED_INPUT = 4_000;
