+++ src/store/hook-debounce.ts
@@ -2,7 +2,7 @@ import fs from 'node:fs';
 import path from 'node:path';
 import crypto from 'node:crypto';
 import { canonicalProjectRoot } from '../core/project-path.js';
-import { NormalizedHostHook } from '../cli/agents/host-hook.js';
+import { NormalizedHostHook } from '../core/host-hook-types.js';
 import { HOOK_CAPTURE_DEBOUNCE_MS, sweepDebounceClaims } from './retention.js';
 
 // The window and the lifetime of the file that enforces it are one decision, so they are
